refactor : 필드 이름 변경, 필드 타입 변경, MockBean 추가 - couponId -> id로 수정 - star count short -> int 수정 - StoreService mockbean 추가

+++ src/test/java/com/prgrms/bdbks/domain/testutil/StarObjectProvider.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class StarObjectProvider {
 
-	public static Star createStar(User user, short count) {
+	public static Star createStar(User user, int count) {
 
 		return Star.builder()
 			.user(user)
